Private methods to bypass buggy namedGroups cache
PL/Java builds and passes its tests on Java 20 this way.
Of course, this code relies on Java 20 API and can only
be built by bumping <release> to 20 in the pljava-api POM.
To do: make it reflective so it can build on earlier
versions, and apply the workaround only if needed.

Begins to address #435.