Who ever noticed 1-letter identifiers don't work?
The runtime parsing of function AS strings, since 1.6, has been
requiring a Java identifier to be a single javaIdentifierStart plus one
or more javaIdentifierPart, which of course should be zero or more.

The result is an unexpected error if a one-letter identifier appears as
a package-name component or a class or method name.

Addresses #438.