Don't over-narrow permissions fetching jar URL
If an HTTP URL is used, either a SocketPermission or a
URLPermission is supposed to work, but HttpURLConnection's
getPermission() method predates URLPermission and just
returns the other one. That can lead to an empty set of
effective permissions if the policy granted one and the
set is intersected with the other.

Also, for HttpURLConnections, install an Authenticator
that will try to find userinfo in the requesting URL,
if the remote end sends a password authentication challenge.
Beware, though, that sqlj.jar_repository has a jarorigin
column, and URLs used to install jars show up there!

Both of these items are from issue #425.