Simply reject Java with JDK-8309515 bug
Trying to include a workaround is simply too brittle;
static methods that do the right thing in place of
Matcher.{start,end,group} would have to be used not
only in PL/Java's own code, but in anybody else's code
that might use the Lexicals class; such workaround
methods might have to be exposed and clutter the API,
and there would still be no guarantee they'd be used
everywhere needed.

JDK-8309515 has been fixed for Java 21 (in
openjdk/jdk@90027ff) and will presumably be
backpatched to some build of 20. So simply test for
it, and if necessary, advise using a working Java
version.