Add objects to extension before CREATE OR REPLACE
Until postgres/postgres@b9b21ac, CREATE OR REPLACE would silently absorb
the object, if preexisting, into the extension being created. Since that
change, those CREATE OR REPLACE operations now fail if the object exists
but is not yet a member of the extension.

Therefore, ALTER EXTENSION ADD them first.

Because it's possible to be updating from an older PL/Java version
(for example, one without the validator functions), this should be
done in Java code where a failure because the expected object
isn't there yet can be treated as benign.

This special treatment is only needed for the few objects
that get specially CREATE OR REPLACEd within the Java code.
The rest can be handled with ALTER EXTENSION ADD commands
in the extension script, as for any other extension.

In passing, also use the tableoid system column to simplify
an older query slightly.