New ACL rights appear in PG 15 and in 16
15 introduces two rights that can be granted on GUC settings
(where no such thing was formerly possible). postgres/postgres@a0ffa88

16 adds one new right included in OnClass. postgres/postgres@60684dd

So N_ACL_RIGHTS can't just be CONFIRMCONSTed; it has to be picked up
as a PG-dependent constant.

For now, assume that the unused bits (in PG versions that
don't use them) will be zero, so no special effort to mask
them off will be needed.

While this commit adds a Grant.OnSetting interface, it does
not add any new interface modeling a setting and its grants.
That is left for another day.

Also left for later is what the methods for the new rights
ought to do in a PG version that doesn't have them. As is,
they will just return false. They could be made to do something
else, like throw an exception for something unsupported, though
that would be a bit more work.