Consolidate the temporary test jig methods
As long as there is a dedicated SlotTester interface,
don't clutter TupleTableSlot with a temporary adapterPlease
method.

For a bit of sanity, make adapterPlease care a little more
about what implementation classes you might name. Have it
expect a class that implements SlotTester.Visible. That can
go away, of course, as soon as there's a proper adapter
manager.