Additions to AbstractType
Add an AbstractType.MultiArray to provide manipulations
of Java multidimensional arrays, restricted to those
matching the shapes of PostgreSQL arrays (non-'jagged',
all arrays at a given dimension having the same size,
no intermediate arrays missing).

In the expected application, a MultiArray will be
constructed, with some number of dimensions, over an
element type returned by some selected Adapter. The array's
type will therefore be that element type with some number
of bracket-pairs added. The possibility must be allowed that
an Adapter for a PostgreSQL element type also produces a
Java array of something. So, the number of bracket pairs
in the resulting array's Java type will be the sum of the
dimensions of the multi-array and those of the possibly-
array type produced by the element adapter.

Strictly, only the levels added by the multi-array must
conform to the PostgreSQL shape constraints (fully
allocated, no jagged sizes). If the element adapter is
producing Java array results, those may include nulls,
be variable in size, whatever that adapter wants to do.

That is, if the final type is [[[[[D, that could be a
five-dimension array of double, or a three-dimension
array where the element adapter returns a [[D. The
MultiArray 'part' is a prefix of the resulting type
(when spelled as above, or in the order of dereferencing
elements; a suffix when spelled double[][][][][]).

When a MultiArray is converted to a Sized, only as many
sizes are supplied as its number of prefix dimensions.
When that is converted to an Allocated, only that much
of it is allocated; the consumer's next task will be
to iterate over the arrays at its last level, populating
those with whatever the element adapter returns.

An iterator is provided to produce those arrays in the
right order. Each will be a one-dimensional array of
whatever type the element adapter produces, be that a
reference or primitive type. By iterating over the
final-level arrays rather than individual elements,
the hottest work can be done in a simple open-coded
loop that may get well optimized; also, further
optimization may be possible under detectable conditions
(known primitive adapter, fixed size, no nulls), such
as direct copy over the array.