Add ArrayBuilder.build() and a suitable service
ArrayAdapter is in the internal module, so define a service
within Adapter (in the API module) to request the construction
of an ArrayAdapter with a specialized Contract, derived from
the MultiArray computed here.

How to handle the resulting type will be tricky. Adapter's
existing rules for teasing it out of the contract's type
arguments won't be enough. There should be some way for
this code to explicitly supply the Type.

There is already a 'witness' constructor parameter of type
Type (at Adapter's level, anyway, if not yet subclasses).
And Type is an unsealed interface. A custom TypeWrapper
class, only instantiable here, can be supplied as that
argument, passed up through the constructors, and
promptly unwrapped to do the right thing. Slightly
distasteful, but avoids a slew of new constructors.

For consistency, tweak the service loading in
CatalogObject to match the slightly more careful
usage here.