Add PG_VERSION_NUM to ModelConstants
For supporting older PG versions, there may be no alternative
to some version-dependent logic on the Java side.

This might be freely used in the implementation module, and
exposed in the API module (there could be user code that also
wants to know), but more cautiously used within the API module
itself, as it can only be available via the factory service
from the implementation.

It seems appropriate to have the version number at the very
top of the constants, even though it means renumbering them all.