Let formObjectId accept a version predicate
For compatibility with PG versions older than what is modeled,
an object will still be constructed (the API interfaces cannot
very well change what statics they export), but with InvalidOid
instead of the oid assigned to that object in the later
PostgreSQL version. So, the object can be retrieved from
the API static, but isValid() will be false, which is about
the right thing to see happen in a PG version that hasn't got
such an object.

Also have OidAdapter.Addressed avoid claiming it canFetch
a type with isValid() false; this is the adapter most likely
to be affected by PG types (of the reg... variety) absent in
older PG versions. (An argument could be made that Adapter
should be testing for this, universally.)