Enter TargetList and its subinterface, Projection
The earlier, just-get-something-going API on TupleTableSlot requires the
caller to get one value at a time by passing an Attribute and an Adapter
to a get() method. That has several practical disadvantages.

Passing an Adapter to each get() method is essentially unavoidable; any
API that wants to use Java genericity so the values' types can be known
at compile time will need to be designed so that the (typed) Adapter the
calling code has selected appears as an argument to a method.

That creates the possibility, unlikely though it may be, that the calling
code could supply different Adapters on different calls, even for the
same attribute. There is non-trivial work to check a supplied Adapter's
canFetch() against the Attribute / RegType being fetched, and select the
corresponding Datum.Accessor. Likewise, the Attribute used in the call
must be checked for belonging to the slot's TupleDescriptor.

To repeat those checks on every get() call would be to spend cycles
needed for a highly unusual use case, while in the common case the
caller will be choosing one Attribute and one Adapter per result column
and using those consistently.

Of course the work can be cached and reused as long as the same
Attribute and Adapter are seen, but the question arises where to keep
the cache. For now, it is with the TupleTableSlot.

That works well enough when PostgreSQL supplies a set of tuples in
a form, like SPITupleTable, in which they are bare, and can be
associated one-by-one with a single persistent TupleTableSlot.

However, some other PostgreSQL APIs exist, and may be worth supporting
at some point, in which each tuple arrives in its own (native PG)
TupleTableSlot. How to handle that will present questions for another
time (creating a fresh Java TupleTableSlot for each native PG one sounds
unappealing), but it seems worthwhile to hoist the aforementioned
checking and caching out of TupleTableSlot, and into some other class
dedicated to selecting the attributes wanted, once, in advance of
processing the tuples.

Such a class can also lighten the weight of TupleDescriptor. In an API
based on one-at-a-time get() calls, which might identify an attribute by
name, there is strong incentive for TupleDescriptor to carry a hashtable
or the like, just to reduce the cost of repeated name lookups. Again,
the cost incurred might be needed in unusual cases, even though in the
usual case the same programmer writing the Java code also wrote the
query, and probably wants most of those columns in mostly that order.
A completely simpleminded method to select several Attributes at once
can offer linear behavior for that case without lugging extra data
structures around.

To that end, enter TargetList and its subinterface, Projection. As in
the algebraic usage, Projection is a sequence of Attribute with none
appearing more than once. TupleDescriptor is-a Projection, of course,
and the only way to obtain another Projection is from an existing one
by reordering its attributes or eliminating some. Its project(...)
methods allow doing that by name or by index. As a result, the
"attributes all belong to the original descriptor" condition is
satisfied by construction.

The superinterface TargetList relaxes the nonrepetition condition;
attributes can appear more than once in a TargetList. That should sound
less efficient than simply mentioning the attribute once and letting the
Java code copy the fetched value, but there may be cases where it is
useful. One would be when the Java code wants different representations
of the PG value, produced by different Adapters. Another can be when the
Java representation is a one-use-only class like SQLXML.

Once a TupleDescriptor has been molded into the Projection (or, more
generally, TargetList) with just the attributes the Java code wants to
retrieve, the TargetList can be applied over a List of TupleTableSlot
at a time, retrieving rows with methods that take a number of Adapters
at once, and a lambda with corresponding parameters, taking their
compile-time types from the Adapters. Functional interfaces of some
likely lengths are provided, and a Cursor object is supplied, both to
permit iterating over the rows and to permit the currying of lambda
bodies that fit the various available apply(...) methods, in order to
fit a TargetList with any number and types of columns.

This commit does not apply any of the optimizations contemplated above;
the implementation, for now, calls the same old heavyweight get methods
on TupleTableSlot. It will, however, pave the way, and those get methods
should eventually be deprecated.