Start on unwrapping ResultSet as Portal
The preliminary SlotTester.test method taking just a simple string query
is too limiting for much useful testing. The ability to construct a query
with parameters may be sorely missed.

In a transitional period while the new API is incomplete (and read-only)
and the legacy JDBC implementation is still present, it would be useful
to allow queries (including parameterized ones) to be issued using the
legacy JDBC mechanisms, and then the results processed with the new API.

To that end, allow SlotTester.unwrapAsPortal to be applied to the
ResultSet obtained from a JDBC query (as long as the ResultSet has not
been used to fetch anything yet). The Portal instance will then support
retrieving the tuples in the new API.