Decree: applyOver methods can throw SQLException
... in addition to whatever the lambda bodies appear to throw.
Some SQLExceptions, wrapped in the unchecked AdapterException,
may not be unwrapped until caught in applyOver.

This is probably not the final story for exception handling.
For one thing, it means code in the lambda bodies could have
catch (SQLException ...) and still miss these, which is weird.
But it is also unpleasant to write functional-style Java while
dealing with checked exceptions.

It also remains to be decided whether the Java-centric,
JDBC-defined, checked SQLException hierarchy is the right fit
for a more JVM-language-agnostic, PostgreSQL-specific PL/Java
API. And the "thoughts on logging" wiki topic has thoughts
impinging on that.

There are nice things about the SQLException hierarchy.
That it carries SQLSTATE codes, and has an exception
hierarchy with a defined correspondence to them, is nice.
And no API as imagined in "thoughts on logging" exists yet,
though things could be much simpler if it did.

It is often right at the point of the throw where code knows
what the most appropriate SQLSTATE code should be. Maybe an
unchecked exception hierarchy paralleling SQLException would
be worthwhile.

For now, this peculiar hack.