Deprecate old hasty TupleDescriptor methods
Deprecate attributes() because TupleDescriptor already is-a Projection,
which is-a List<Attribute>, so it hardly needs a method that only
returns its own receiver.

Deprecate the two get() methods that do lookups by name, in favor of the
methods of Projection, which can look up multiple attributes by name at
one time, addressing the inefficiency of repeated one-name lookups. When
these methods have gone, there should be little reason for TupleDescImpl
to be hauling around a hashtable.

Nothing is wrong with sqlGet, but hoist it out to a default method on
TargetList.

And deal with the resulting fallout ....