Add test for CatalogObject class initializations
Now that they project all the expected attribute names in advance,
they will break more dependably on PostgreSQL versions that don't have
the same attributes, but still only when methods requiring holder class
initialization are used, so it is good to have a test that does that.

Not departing from the task at hand, add an AttNames utility class in
CatalogObjectImpl to simplify assigning these Attribute static finals
conditionally, based on the PG version, and use it in RegTypeImpl and
RegProcedureImpl to get them working back to PG 13 (as far back as this
can otherwise be built without other patches).