Migrate SQLXMLImpl.Readable, VarlenaWrapper->Datum
The only mentions of VarlenaWrapper remaining in SQLXMLImpl are the ones
for the to-PostgreSQL direction (for which there isn't any support in
Datum or TupleTableSlot yet). It should, in principle, be possible now
to supply it with Datum.Input instances that are not VarlenaWrapper
instances.

Some methods that weren't exposed as Datum API are now there where they
belong.