Add a set-returning function example
When a Portal allocates a TupleTableSlot, it had better do so
in a context as enduring as the Portal itself, and not blindly
in the current context, which is likely the SPI context that
will be reset for each call in the SFRM_ValuePerCall protocol.

For now, just allocate it in the Portal's portalContext
directly. This is arguably a little incestuous, but the fewest
lines of code to change at the moment to get it working (as
MemoryContextImpl.java doesn't yet expose any methods for
creating and managing a new context), and it has the right
lifespan.

That is, it has the right lifespan unless a holdable portal
is created, and persisted; then stuff should be in its
holdContext (or some other context at least as long-lived).
So everything is still definitely stopgap in this commit.
TupleTableSlotImpl still isn't aware of its memory context
or lifespan, even.