A reference NUMERIC -> BigDecimal implementation
BigDecimal is nearly a fit, as things in the standard Java
library go, but it cannot handle NaN or +/- infinity. At any
rate, even if this reference implementation of a contract
isn't complete with respect to those values, it could easily
be wrapped by another contract that could return a subclass
of BigDecimal, or a BigDecimal-or-Something ... or some other
type that is easily constructed from a BigDecimal.

Because PostgreSQL's numeric.c contains the magic numbers, and
they aren't exposed in .h files, a regression test is critical
to making sure these bits don't rot.

In passing, relax the obsessive schema-qualification in
TupleTableSlotTest's SQLActions. These are (a) examples, and
(b) when run as CI tests, running in a fresh installation,
and not all other examples are obsessively schema-qualified,
and this one didn't have the operators qualified anyway, and
doing *that* makes an example just plain hard to read.

# Conflicts:
#	pljava/src/main/java/org/postgresql/pljava/pg/adt/NumericAdapter.java