Improve NEWLINE pattern
Old one could fail depending on previous matching activity, possibly
because of the use of \G (which I should have explained better in a
comment, back when I thought I knew why I was doing it). The documented
behavior of ^ $ and \z and reluctant quantifiers make for a simpler
and more dependable version.

Addresses issue #455.