Pick up SIZEOF_INT via ModelConstants
In backporting, sometimes the git history shows that something
has always had the same type, but the type was plain int rather
than an explicit-width type. So, for such things, there is no
need for a plethora of SIZEOF_FOO constants, but SIZEOF_INT may
be generally useful to detect if a platform has a surprising
value for that width.