Reported changes needed for TBase
TBase diverges from PG 10. Reportedly the offsets of
some fields in struct HeapTupleHeaderData are different.

This commit demonstrates how to accommodate
such a change. A constant deemed to have been stable over
many PG versions might simply be defined in the Java code
and use a CONFIRM... macro in ModelConstants.c to detect the
unexpected case where it changes.

If a constant turns out to be less constant than that, then
the approach has to change. That means deleting the simple
Java definition of the constant (here found in
TupleTableSlotImpl.java) and the matching CONFIRMOFFSET
in ModelConstants.c, and then adding four things, three
of them in ModelConstants.java: 1) assigned IDX_...
indices for new constant slots; 2) public static final
declarations for the constants; 3) statements within the
static initializer that assign the values with calls of
n.get().

To avoid repeated merge conflicts with ongoing development,
the indices assigned for these new constants can leave a
comfortable gap (say, start at 1000) above the indices
already used, and use n.gap() to account for it.

The fourth thing to add will be lines in ModelConstants.c
that store the needed values into the assigned slots.

Because TupleTableSlotImpl.java used to define those
offsets locally, and now doesn't, it could be necessary
to add a static import from ModelConstants.java. But
in the case of TupleTableSlotImpl, there is already
an import static ModelConstants.* so no change is
needed there.

The nonstandard CREATE TABLE ... DISTRIBUTE BY
REPLICATION of course make this patch a bit tricky
to test on vanilla PostgreSQL.