Add a legacy-API example that won't work yet
This commit illustrates what could be needed to get 2D primitive
array function return to work under the legacy C/JDBC API.

This example generates a usable CREATE FUNCTION that PL/Java
will attempt to load. The validator rejects it with "cache lookup
failed for type 0". With SET check_function_bodies TO off; to
skip the validator, the function can be created. Then the "cache
lookup failed for type 0" will happen on the first attempted use.

What is happening at this stage: after Function.storeToNonUDT,
the Type object for double[] has been proposed for the return
type, based on the SQL declaration. But the CREATE FUNCTION
Java signature has explicitReturnType of double[][], unequal to
the proposed type, so _reconcileTypes(..., -2) will be called.

In _reconcileTypes, Type_fromJavaName(..., "double[][]") is
called, and successfully returns replType based on the existing
Type for double[] and createArrayType. The replType returned
identifies itself as type.double[][].

In the next step of _reconcileTypes, this new Type is asked,
via its _Array_canReplaceType method, whether it can replace
the originally proposed Type. It answers no, so _reconcileTypes
tries Type_getCoerceOut from one to the other, and that is
what fails, looking up InvalidOid in the cache.

This suggests a few possible ways forward, with slightly
different implications. In no particular order:

1. Change what Type_fromJavaName returns. It currently
   knows it can try removing one [] from the end at a time,
   call itself recursively, and apply createArrayType to
   that result. For double[][] it does this twice, finds
   double, whose overriding _double_createArrayType returns
   a specialized array Type with its own coercers. When the
   second createArrayType is applied to that, the non-
   overridden implementation creates a generic Array with
   the double[] Type as its element type.

   But while double[] and double[][] are different types in
   Java, they are the same type in PG. What if:

   1a. Type_fromJavaName knew it could remove multiple ([])+
       from the end, find the existing type in a single
       recursive step, and apply createArrayType once to
       that, no matter the number of []? It would, therefore,
       return the Type for double[], whose canReplaceType
       would trivially return true (it's the same Type as
       the proposed one). _double_coerceObject would need
       to be able to handle multiple dimensions.

   1b. With Type_fromJavaName unchanged, _double_createArrayType
       created a new Type with not only overridden coercers,
       but also an overridden _createArrayType that would
       return self on later calls? The result should be the
       same as 1a.

   1c. Hmm, given that thing[] and thing[][] are always the
       same PG Type, perhaps every array Type should always
       have its own _Array_createArrayType that just returns
       self.

2. Leave the behavior of Type_fromJavaType unchanged, and
   change the behavior of _Array_canReplaceType. In this case,
   Type_fromJavaType would still return a new Array reporting
   the Type for double[] as its element type, but its
   _canReplaceType would answer yes it can replace its own
   element type as long as that element type is also an
   array. In this case, it would be _Array_coerceObject that
   needs to be able to handle multiple dimensions,
   delegating (for the lowest dimension) to the _coerceObject
   of the Type it is based on. This approach be more general.