Try the array type of an array type being self.
This is idea 1 (specifically 1c) from the preceding commit
message. As expected, it results in the final Type used
for the function return being the existing type.double[]
and the function is successfully called. Nothing in
getMethodHandle needs changing.

Because the function returns null, it works fine. Making
it work for non-null values is left as an exercise. :)
The exercise in this case would take place in the
_coerceObject method of Double.c.