Try Array type replacing element type if array
This is idea 2 from the preceding commit message.

That looked easy enough, but the simple reference equality
test Type_getElementType(self) == other fails, because it
turns out there can be more than one Type object (even
pointing to different TypeClass objects!) for the same array
type, and that turns out to be because of a change three
years ago in 19afa36, whose commit message should be read
in full for context.

Back to the drawing board.