Work around the Array type nonuniqueness headache
This will work for 2D arrays but no higher. Extending it
to more dimensions would be an exercise.

As expected, it results in the final Type used for the
function return being an Array named type.double[][]
and the function is successfully called. Nothing in
getMethodHandle needs changing.

Because the function returns null, it works fine. Making
it work for non-null values is left as an exercise. :)
The exercise in this case would take place in the
_coerceObject method of Array.c (which would need, at
some lower dimension, to delegate to the _coerceObject
method of the element type).