Let stateMachine states make relative moves
It would be nice to have a little library of common
stateMachine states, heretofore difficult because the
next state must be indicated by returning an absolute
state number. So, use one of the otherwise-unused parameters
of the InvocationHandler functional interface to supply
a state with its own state number, so it can make relative
moves.

To start the library, here is NOTHING_OR_PGJDBC_ZERO_COUNT,
because the PGJDBC and pgjdbc-ng drivers differ in handling
a utility statement with no result (CREATE EXTENSION, that sort
of thing). In pgjdbc-ng, there is no result, while in PGJDBC,
there is a zero row count, as would be seen for a DML statement
that had not updated anything.

So, NOTHING_OR_PGJDBC_ZERO_COUNT simply moves to the numerically
next state, after consuming (if the driver is pgjdbc-ng) nothing,
or (if the driver is PGJDBC) a zero row count.