Close a Node's Connections when stopping it
This is all about convenience when using JShell interactively,
where try-with-resources gets in the way. The VM shutdown hooks
to stop and clean up the Node work well, but if any Connection
is open at the time, the type of shutdown signalled to the
server may not cause it to be closed, leading to a hang. The
Java Process API does not expose enough types of OS signals
to easily request a faster shutdown. (When the node is using
pg_ctl, that's an option, but not when signaling the server
directly.)

So, have each Node remember the Connections it makes, in a
WeakHashMap. They'll be removed from the map once unreachable
(and if that happens before they're closed, both drivers
include a cleaner task that will eventually close them; there
may be a bit of delay, but that's better than the former hang).
When stopping a Node, close whatever Connections are still
around in the map.