Update the documentation for the Node test harness
Update the docs chiefly to reflect that either PGJDBC or pgjdbc-ng
may be used as the JDBC driver, and how to accommodate the slight
differences between the two. Numerous other updates, including to more
consistently observe the conventional javadoc style where lead sentences
are third-person rather than second.

In the course of better documenting when tweaks are applied,
caught one place where forWindowsCRuntime would not be applied
on Windows. (The known arguments being supplied right there in
the code could squeak by without it, but who knows what could be
added in a tweak?)