Quit using paired entry points to pass 'trusted'
The legacy dispatcher has needed to look up the language in
the catalog anyway, and so will the new dispatcher, and the
only use formerly being made of the 'trusted' value gleaned
from the entry point, and assiduously passed along, was in an
assertion that it wasn't different from what we found in the
catalog. (That, and bifurcating the "save the first oid that
refers to PL/Java" logic into two cases, which were everywhere
used in the pattern "nothing saved in this one? ok try that one".)

There still are the paired entry points, they just don't do
anything different. Changing the SQL declarations can be for
another day.