Tidy some uses of currentInvocation->function
A handful of functions to get information about the Function
pointed to by the current Invocation were all in Function.c,
but for one living all alone in Invocation.c.

Centralize them all in Function, to simplify adding a new
dispatcher that won't be using C Function structs for its
dispatched routines.