check_function_bodies readable from Java
This commit introduces a ByteBuffer[] _window method in Backend
that can be used to window various miscellaneous PG globals that
don't obviously belong someplace else. (The technique is already
used in MemoryContextImpl and ResourceOwnerImpl, for example, for
several globals in obvious groupings, but there may as well just
be one place consolidating uses for less closely-related things.)

The first such global is check_function_bodies. There may be
several existing JNI downcalls from Backend that could use this
technique instead, as a future opportunity for tidying.