Sundry fixes to CatalogObject implmementations
Add a few casts to CatalogObjectImpl in key places to avoid trying to
interoperate with outside implementations of CatalogObject interfaces,
and improve an exception message that was very preliminary.

The non-private-ness of one CatalogObjectImpl.Addressed constructor was
a crutch while there were subclass implementations partially
implemented.

Narrow an overbroad SuppressWarnings("unchecked").

Have CatalogObjectImpl.toString report the name of the API interface,
not the internal class.

Fix a thinko in AttributeImpl.Transient.equals, and watch for edge cases
in RegType/RegClass tupleDescriptor methods (for example, any regular
relation has a row type, but an index or toast relation does not).

TRIGGER will be a useful type to have around, and the equality of
SIZEOF_Oid and Integer.BYTES ought to be asserted at least somewhere.