Add Identifier.Simple.None
The PostgreSQL catalogs can contain empty strings in some
contexts where a name might not be provided (for example, when
pg_proc.proargnames is nonnull because some parameters have names,
but not all of them do). So let Identifier.Simple.fromCatalog
(but not other methods) accept an empty string, returning
the None instance.