Add TupleDescImpl.synthesizeDescriptor
For a modern dispatcher to PL/Java-based languages, a consistent
representation for the number, names, and types of incoming
parameters and expected results will be wanted. TupleDescriptor
fits the bill. For incoming parameters, the catalogs identify
the names and types (no typmods), and this method will serve
to gin up an ephemeral TupleDescriptor based on those.

PostgreSQL already offers some funcapi methods for getting a
tuple descriptor for the expected outputs, but only when the
routine is really expected to return a composite. For functions
returning a non-composite PostgreSQL type, we will still want
to synthesize a one-attribute TupleDescriptor of that expected
type, so that Java code will always consistently produce a result
by storing something into a TupleTableSlot. (It will be the
dispatcher's job, then, in the non-composite case, to grab the
value from the one-column TupleTableSlot and return it.)