More invalidation: two more classes, and TupleDesc
RegProcedure and ProceduralLanguage are two more object
classes we'd like to cache as persistently as practical,
so it is worth using two more callback slots to be able to
invalidate those selectively.

Renames Factory.invalidateType to syscacheInvalidate with a
cacheId parameter, as suggested in 5adf2c8. Many other
optimization opportunities suggested in that commit still left
on the table.

The precedent set back in 2e74a6b of final SwitchPoint[] foo
= new SwitchPoint[] { new SwitchPoint() }, and then ignored
in 5adf2c8, seems to be the right one after all. After the
initial construction, it is rather tidy that compute methods
run on the PG thread, invalidations come from the PG thread,
and SwitchPoints impose some order on who sees what when.
It's that initial instantiation, meant as a cheap 'whenever
mentioned' operation, where the final-field semantics are
of help. RegType therefore also fixed to match.

Also add tests in the CatalogObjects example for invalidation
of the four object classes now expected to support it.

TupleDescImpl now also has a notion of invalidation, which can only
happen to the Cataloged flavor, and happens upon invalidation of the
corresponding RegClass. Unlike catalog objects, a TupleDescriptor
after invalidation won't magically have updated values; it will just
throw an informative IllegalStateException.