Flesh out the RegProcedure.Memo idea
The vague early idea that a RegProcedure ought to carry a memo still
seems useful, and just how it should be useful is becoming clearer.

There may be no need for user code to apply any memos; remove the
apply() method from the API for now. Also, while RegProcedure declares
the Memo interface, the subinterface PLJavaBased seems more at home in
ProceduralLanguage, where Handler, InlineHandler, and Validator already
are.

The two memo subtypes Validator and PLJavaBased will clearly be of use.
When a RegProcedure p is determined to be a validator, it can be given a
Validator memo with a direct reference to the ProceduralLanguage it is
the validator for (contrast p.language() which is the ProceduralLanguage
that p is implemented in, and in this design will have to be PL/Java's
handler 'language'). Likewise, a PLJavaBased memo can be attached to a
RegProcedure p when p is determined to be implemented in some PL/Java-
based language pl. The memo does not need to hold a reference to pl--
p.language() is perfectly suited to this case--but it can factor out
some complexity of invalidation, and also play a foreseeable API role,
with methods exposing PL/Java-specific information useful to a handler
implementation that goes beyond what every RegProcedure exposes from the
catalogs.

Because CatalogObjects are weakly cached, they can go away when you're
not looking, even without an invalidation notice from PostgreSQL. To be
useful for caching related information, objects that are of interest
ought to be kept live. A static set s_plJavaHandlers can serve as a
root, holding references to instances of the "PL/Java handler language"
(only one such instance is envisioned, but of course PL/Java supports
the idea of language aliases). Each "handler language" instance hl can
have a LanguageSet holding references to those procedural languages
whose validators are implemented in hl. And each of those language
instances can have a RoutineSet holding references to RegProcedure
instances dependent on that language. Thus are the RegProcedures and
ProceduralLanguages of interest kept live.

Those links have to be culled when invalidations happen. An invalidated
"handler language" should remove itself from s_plJavaHandlers and
propagate the invalidation to its dependent languages; an invalidated
ordinary language should remove itself from the dependent-languages set
of its associated "handler language" (for an ordinary language l, that
is l.validator().language()) and propagate the invalidation to its
dependent RegProcedures. An invalidated ordinary PL/Java-based
RegProcedure should remove itself from its ProceduralLanguage's
RoutineSet, while an invalidated RegProcedure that is a validator of
some language needs to invalidate that language.

The different invalidation behavior for a RegProcedure, determined by
whether it is a validator or an ordinary PL/Java-based routine, can be
handled neatly by delegating to whatever memo (Validator or PLJavaBased)
it carries. And of course every memo, on invalidation, removes itself
from its carrying RegProcedure.

To be clear, all of this is determined lazily: no RegProcedure is known
to be a PL/Java-based language or a validator until it is encountered in
that role while PL/Java is dispatching a call. The attaching of memos
and entering into sets happens then. The two actions should be regarded
as duals; the reference held in some set to a dependent object and the
memo on that object should both exist, or not.

ProceduralLanguageImpl adds methods isPLJavaBased and isPLJavaHandler
that a dispatcher will be able to to call when it needs to make the
(rather fiddly) sanity checks that a language instance is set up in the
expected way. These methods, when returning true, take care of adding
the language into the appropriate parent set (s_plJavaHandlers for a
handler language, or the parent handler language's LanguageSet for an
ordinary language). However, it will be the dispatcher's job, when these
methods return true, to make any links involving a RegProcedure and
attach an appropriate memo to it.

Here decreed in these isPLJavaBased / isPLJavaHandler methods is
that the C entry points for the new dispatcher will be named
pljavaDispatchRoutine, pljavaDispatchInline, and
pljavaDispatchValidator. A "PL/Java handler" language will be one
declared with no inline handler, and with both its call handler
and its validator handler pointed at (different SQL function
overloads with) the C entry point pljavaDispatchValidator.